backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Task Queue
celery==5.3.4
redis==5.0.1

# HTTP Clients
httpx==0.25.1
aiohttp==3.9.1

# Stream Processing
python-ffmpeg==2.0.12

# String Matching
rapidfuzz==3.5.2

# XML/EPG Processing
lxml==4.9.3
xmltodict==0.13.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytz==2023.3
schedule==1.2.0

# Image Processing (for logo comparison)
pillow==10.1.0
imagehash==4.3.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
